PNGExporter: fix IDAT zlib writer
This fixes the IDAT zlib writer, which was breaking (but masked in
testing as our image was not big enough to trigger the issues).

Specifically, we need to keep track of bytes written in their
*uncompressed* form, and then flush ourselves when we hit a particular
threshold. We then check to see if we are still below the minimum
remaining bytes, and if we are, write that IDAT chunk out, and reset the
buffer.

I think there could be some optimizations in how this is done (namely
code cleanup), but this is good enough for now. Trying to move the
writing to a dedicated struct did not necessarily go too well (probably
something going out of scope or not being mutable, and ultimately
causing NoSpaceLeft errors).