spec: current final word on spec tests
This is the final code cleanup for spec tests (for now). Unfortunately,
since Zig does not allow test generation right (e.g. something like
t.Run in Go), we need to copy out each test into its own block.
Similarly, we are doing generation of the spec/golden files "manually"
for now as well (this might change, but the code is at least clear in
intent).