pattern: simplify OpaquePattern, make it a value vs reference
After thinking about this more, I've decided to simplify OpaquePattern
for now. We don't need the memory management part yet, and I'm not too
sure when we will. Currently, the pattern exists to provide scaffolding
to the concept of an interface that we will need eventually (e.g. when
we allow surfaces to also be presented as patterns), but I'm not too
sure when that will be, and the tagged union structure means that we
don't need to have every implementation be 100% uniform (e.g. one can be
a value, another can be a reference).