bresenham: draw at end of loop instead of the start
This should catch situations where there's still a pixel to draw after
iteration (on the opposite axis of what we're iterating along).

This removes gaps on beziers, although the result is still pretty
jagged. This is expected due to how the bezier is broken up into lines.
This is something I'll continue to review, but for now, this is good
enough to move into thickness and caps.

I did also attempt the DDA variant, but there were still gaps and even
worse it was drawing pixels in places where it didn't need to. I'm
guessing it needed tweaking, but the gains were marginal over Bresenham.