path: line iteration fixes
Turns out some of the gap issues were caused by off-by-one errors in
some of our for loops, fixed by making sure we add + 1 to every time we
use for to iterate to an end co-ordinate.

This actually proper fixes the Bresenham stroking so that there are no
more gaps, so I've reverted the pixel drawing back to the start of the
loop.