path: refactor, introduce transformer architecture
This moves the path functionality to a sub-package, and introduces a new
fill-and-transformer architecture. This is designed to make it easier to
conceptualize more advanced drawing operations, like stroking,
particularly with a thickness.

This also removes the Bresenham stroker for the time being and also
disables the stroking tests. Eventually, I'd like to (possibly)
re-introduce the Bresenham bits as a hairline stroker (for widths of 1
pixel or smaller). For widths larger than 1, we actually need to build
a more complex polygon which is then filled, hence the need for the new
architecture - this architecture will transform stroked paths into the
proper series of line nodes which can then be plotted appropriately.

The concept behind the transformer can be seen in the
spline_transformer.zig and the new handler for curve_to in filler.zig.
Here, we transform the bezier into line_to nodes and then just plot them
directly.