polygon: allow for multiple subpaths in a single fill operation
This fixes path fill so that it can draw multiple subpaths in a single
fill operation.

This is done for now by simply just using multiple Polygons within the
new PolygonList struct, which wraps the plot and edge-finding operations
so that they're multiple-polygon aware.

This also removes the old edge functionality from Polygon (was needed
for Bresenham, but will not be needed for the time being).