path: new stroker scaffolding
This adds scaffolding for the new stroker, with initial support for
single-point lines.

The stroker works by transforming a path to be stroked into a fillable
path with the specified thickness. Caps are currently butt-style caps
(the line ends on exactly the points defined in the path).

A test is included that ensures that we are seeing the points laid out
correctly regardless of the direction of the lines, e.g., we draw up to
down, down to up, left to right, right to left, and combinations
thereof.