stroke: much better (and more complete) miter logic
This should hopefully put us on the right path for a "complete" miter
join situation (save adding limits).

This introduces "faces" and specifically denotes different kinds of
lines (horizontal, vertical, and diagonal). We define very specific
behavior for each line and process join rules different for each type.

We additionally compute line offsets now off of atan2, which saves a lot
of headache WRT determining correct direction for calculating line ends.

This should set the stage for moving on to path close and curve
calculations, which should then complete the initial stroke
implementation. Note that this initial implementation will not work with
overlapping lines since we are still filling odd-even versus non-zero,
but that will come shortly after.

One final note: we have added thickness to the API for DrawContext.