stroke: add support for close_path
This adds close_path support for stroking, allowing us to add back
several of the closed stroke tests - triangle, square, and trapezoid
have all been validated as functioning correctly now, as seen by their
images.

Star is *almost* working - we now need to add support for non-zero fill
and ensure that stroking uses that for fill operations. This can be seen
in the inner joints of the star - this is where the path segments
overlap with each other during the tracing process.