path: add non-zero fill
This introduces non-zero fill for paths, which ensures that complex
paths that overlap or intersect will be filled according to the
directions of the lines within the paths, versus simply by counting the
odd-even number of edges along the path.

This fixes gaps in strokes that do overlap, since we now enforce filling
strokes with the non-zero rule (without any way to change it). This can
be seen in the star stroke test.

The process could probably be optimized a bit given that this continues
to increase the amount of memory allocations we need to do per scanline.
A couple of things we could do: ensure edge list is handled directly in
fill (versus the outer edgesForY), and switch to an ArenaAllocator to
speed up the downstream ArrayList allocations (this latter one I need to
explore though, to ensure that we don't just continually allocate memory
for ArrayLists that are only needed for a short period of time).