path: add stroke bevel join, and miter limit
This adds bevel join and miter limits to stroking operations.

Bevel joins simply join the two unjoined outer line points together,
giving a cut-off look.

Miter limits allow for the drawing of a bevel when a drawn miter would
normally be too long. This is defined as miter_length / thickness, with
the miter length referring to the whole of the miter (note: note the
middle of the line).