path: add curve stroking
This adds curve stroking! The idea is similar to how we handle curves
during fill - decompose them to line nodes and then plot the lines.

Some oddities I found:

* Looks like our spline transformer is adding the initial point as a
  line, redundantly. I've set a guard to prevent the spurious addition
  within the transformer for now, although I feel like there could be a
  better way.

* There will be situations in round joins where we end up with zero
  points - in this situation, bevel instead. This will likely happen a
  LOT with regular curves, so I'm wondering if there's a better way, but
  this way works for now and does not break any other tests.

This now adds handling for EVERY path operation! We are close to
completion of stroking completely - to finish things off, we need to
work on caps, which hopefully should be easy now that we have the pen
written for handling round caps. Butt caps are our default, and square
caps should not be too hard.