path: refactor stroke face code
This takes the code for Face in the stroke transformer, moves it to its
own file, and removes the tagged union.

This was a long time coming - it was pretty clear a bit ago to me that a
tagged union was not the right solution here given the large number of
accessors with little unique functionality between subtypes.

To still allow for special handling of different types of intersections,
we still maintain a type tag for each face, but just as a field, and
switching when we need to for each intersection type.