path: add stroke line caps
This adds line caps to strokes - all three styles (two new technically):
butt caps, where the line stops at the endpoints (already implemented by
default), square caps, where the line extends from the butt at width / 2
(making it square around the middle), and round, which utilizes the pen
functionality that we added for round joins to create a 180 degree
rounding around the end.

We include some hopefully extensive testing for these cases that
illustrates the behavior (including even demonstrating the power of the
path engine by drawing the main strokes, and then a "hairline" stroke
over the main lines to validate the caps and how they relate to the
lines themselves).

This, at long last, completes the initial stroking implementation - yay!
There is probably a LOT of optimization that could be done, but the base
algorithms exist, which which we can build the rest of the library on,
accommodate edge cases, etc.