spec: use butt cap hairline stroke on rounded test
This ensures that the hairline shows up.

This likely showed up with the change of the polygon point finding
algorithm to use rounded values which possibly turned the "round-capped"
hairline on certain co-ordinates to be a single pixel wide.

I have to mull how to fix this; I think time will show if this is
causing any more issues, but this was changed in the first place to
ultimately make specific shapes (namely fills) to align on x-boundaries
consistently with each other between pixelated and non-pixelated
iterations, but this may not be the correct approach.

In the meantime, we currently are not dealing with fully degenerate pens
in the correct way - specifically we account for some degeneracy
scenarios in initialization, but not in verticesForJoin, which is
handled in Cairo. This is where we should probably look to fix this
first.