path: restrict stroke options to minimum width
This enforces a minimum stroke width on the following options for now:

* Line join mode (default .miter)
* Miter limit (default 10.0)
* Line cap mode (default .butt)

For stroke widths smaller than 2.0, the defaults will be used, and any
options for these overridden.

This is only expected to be temporary, and is just designed as a
stop-gap until the path builder can properly handle these when the
stroke width is smaller.

It's also expected that we still will need to have some of these
exceptions, but what will likely happen is that these will be moved to
cover very specific edge cases where they pop up.