path: move fill/stroke to context
This moves the fill and stroke operations to DrawContext.

The rationale: most of the parameters these operations need reside in
DrawContext, and the fill/stroke operations needed to reach back into it
for these details. Having the operations themselves live there should
reduce the complexity in this regard.

In addition to taking the path as parameters here, we also take the
allocator. This should make it pretty clear what is required for the
operations.

We've also made the array lists in the filler/stroker bits
"pass-by-value"; quotes are used here because ArrayList actually stores
a pointer to the underlying slice, so nothing is actually copied other
than capacity and the allocator (which in turn is also just pointers).
Hence, the items in the ArrayList itself are not copied. This also makes
all of these parameters immutable, meaning we can't use any of the
methods on the ArrayList, which should communicate that the passed in
nodes are read-only.