Rename "path" sub-directory to "internal"
This is just to denote internal API, in other words where we're putting
source files that are not a part of the main API (and likely will not be
included directly through z2d.zig, and we will avoid trying to have
included indirectly either).

I'm naming it such for the lack of a better name or organization scheme
that doesn't seem excessive at this point.