General assert cleanup
This moves most of the asserts that were in the codebase to errors.

While they are still edge cases, in the spirit of safety, I don't
necessarily want them to be optimized out. Most of them are state
machine related; in this case they return InvalidState, which we are
going to earmark as an internal error value.

There are a couple of other errors here: InvalidPathData for errors
pertaining to invalid data passed to fill and stroke, and
UnsupportedSurfaceFormat in the PNG writer.

All of these errors will be consolidated/codified in #23.

Fixes #24.