Painter/Context: configurable tolerance
This adds configurable tolerance for our curves and pen. This affects
the following:

* Error tolerance when calculating spline decomposition.
* Minimum arc height when calculating the pen.

The idea is that higher tolerance requires less computation when
plotting these kinds of things (less recursion, less vertices, less
plotting). It also produces "blockier" beziers, which could be desirable
in some scenarios.

There is a warning placed on the setting for using this setting with
rounded joins and caps; this is due to the that pen vertices are static
along the pen's circle. This means that you're really not going to get
what you want, which can be seen in the acceptance tests (which are
deliberately set the way they are to illustrate this).