FillPlotter: fix issue with close_path and degenerate paths
This fixes an issue with FillPlotter where a plotted polygon may not
have been pushed to the final list if the current point was the same as
the initial point.

Also adds tests for cubic beziers and arcs. Note that we don't have
primitives for these right now, but they can be represented with quad
beziers.