StrokePlotter: process joins based on first direction
This processes (most) joins based on the direction of the very first
join processed when stroking a path. This sets the stage for the rest of
the stroke and ensures that the points for the inner and outer joins are
placed in the correct polygon, even when the path of the stroke changes
direction in the middle (i.e., from clockwise to counter-clockwise, or
vice-versa).