Path: add arc support
This adds support for arcs for Paths (drawing of circles, and eventually
ellipses once we add transformations).

Like Cairo, arcs are not added as path primitives themselves. Rather,
the arc function is a helper for finding the best-fit approximation of
the arc using splines within a specified tolerance (can be provided, but
is optional; if not, the default tolerance is used). The arc
approximation logic has been taken from Cairo
(attribution/acknowledgment in file).

The API is similar to Cairo, but the normal/negative methods have been
condensed into one. This may not be set in stone; I'm contemplating
adding shorthand normal/negative/default tolerance methods.

Also, regarding tolerance: the main reason we require this in the method
is that we de-couple Path and Context. As such, we cannot take the
tolerance from the context. This implementation detail is only slightly
inconvenient I think, and there are notes in the method description to
instruct on when it should be changed from the default, and where to find
it in the context. It should be noted that it appears that Cairo itself
only takes the value upon execution of the arc functions, so if you add
arcs, and then change tolerance after, the drift will not be reflected
in the approximations (the splines will have already been written to the
path).