StrokePlotter: handle co-linear joins
Fixes #37

This fixes co-linear joins by just plotting the end point of the inbound
face, regardless of the join type (this case holds for every type as the
angle between the inbound and outbound lines would always be 180
degrees).

Additionally, due to the fact that we can't specifically tell if a
co-linear join is clockwise or counter-clockwise, we defer the
determination of rotational direction until a point that we can actually
determine it. If it never is determined (i.e., the path is essentially
just a straight line), we just assume counter-clockwise (derived from
our calculation of cw = in.slope.compare(out.slope) < 0).