StrokePlotter: new inner join behavior
This fixes inner joins so that they behave like they do in most other
graphics libraries, in that rather than attempting to take the
intersection of the two inside faces/edges of a join, the polygon edges
are actually traced through the endpoints of each line as if they were
unjoined, going through the midpoint.

This creates edges that overlap and give the same effect, while avoiding
edge cases on very acute angles where stokes may actually overlap on the
inside, and as such their inner faces technically do not intersect at
all.