StrokePlotter: some semantics changes, removal of not needed state
We've calculated a single direction for the larger polygon for a long
time now (based on the first join's direction), but we still had
left over state for the start and end directions of a particular
polygon. This removes that and also fixes some semantics in the join
method so that it's clear how the method handles direction.