Face: simplify intersection logic
This removes all of the special-case and intercept-based intersection
logic from Face.zig, replacing it with a more generalized version.

The intersection equation has been taken from Cairo, as they have a good
reduction that allows the intersection to be calculated via first
calculating the y-coordinate, and then calculating the x-coordinate from
the larger y-delta between the inbound and outbound faces. Cairo does
this to prevent division close to zero, and as such it works when one of
the faces represents a horizontal line.

This has also allowed us to remove the special-case initialization as
well, and even with this, none of the acceptance tests have changed, so
any slight differences that may happen due to generalization of offsets
has no ultimate effect.

(Note that it's possible that when we eventually implement
transformations, this could have had an impact, but in that sense, it's
better this change now versus then.)