internal/Painter: AA bounding box rounding fixes
More to come, but this fixes the bounding box for AA paint so that the
offsets and dimensions of the mask properly match that of what the
bounding box would be. This is particularly prescient for situations
where we have negative co-ordinates in device space, where intFromFloat
alone would actually be truncating and rounding towards zero instead of
away from zero, and would have the effect of slight shifting and
shrinking the viewport.