Path: add transformation support
This adds transformation support to Path, and the start of an elliptical
arc test using the new transformation functionality.

This is a rudimentary implementation where to modify the CTM in a path,
you just re-assign it through composition of the existing CTM. Points
added to the path are transformed through the CTM in place at call-time.

Stacking matrices for save/restore operations is an exercise left to the
consumer. I'm debating adding helpers to let you mutate the CTM
in-place, just to keep things less wordy, but aside from this, I'm not
planning on having intricate save/restore Cairo-like functionality. We
have an open TODO for infallibility and adding another ArrayList that
would need allocation seems to be a step in the wrong direction.