internal/Face: add Context transformation matrix support
This adds the Context transformation matrix and effects it on Face for
stroking, allowing line width to be influenced by the rotation and scale
of the CTM set in the context.

Added a test in the ellipse test to demonstrate the effects of this in
that you will see observable expansions and contractions in stroke
thickness on the major and minor axes, respectively.