internal/Face: fix offset calculations to properly transform rotations
In order to properly transform offsets when not on a 90 degree exact
rotation, we need to be able to do it in a pos/neg kind of way, as
opposed to the way where we're doing it right now where we apply
absolute co-ordinate offsets and then offset according to our end
points.

This is basically how Cairo calculates its face offsets. Our concept of
"clockwise" and "counter-clockwise" is inverted, but we just reflect
that in how we define our cw/ccw offsets versus the offset calculations
themselves (for my own sanity really).