internal/Face: fix square caps
Using `@abs` for the offsets shipped is the incorrect move here, what we
want to do is invert the x-axis as that's our offset that is inherently
inverted in our updated base case. This may or may not be what we need for
square caps when applying transformations (needs testing; Cairo for
example does a transformation off of the user vector, but I think we've
effectively done that already during init).