internal/StrokePlotter: support transformations on cap/pen
This properly adjusts square caps and the pen (round caps/joins) to tak
into account the context CTM.

Internally, this also drops the use of un-normalized slopes. This
actually simplifies things for the most part, as our caps and end point
are now just calculated off of the normalized slopes * thickness / 2.

We also now impose a minimum line_width, which should ensure that we
never go into negative values that would cause undefined behavior when
plotting. Somewhat arbitrarily, I've chosen the minimum that can be
represented in Cairo's 24.8 fixed-point format, which is a nice clean
1 / 256 (0.00390625). This actually helped some of the edge cases in
044_line_transforms_smooth more closely match Cairo.