spec: add test for round joins w/transformations
This adds a test for round joins (and round caps) rotated in various
directions.

The alignment is not 100%, which is due to how rotate-about-origin works
and me getting tired of fiddling with offsets. :P However, and to the
points in the comments, the image renders the same with the equivalent
Cairo code, so I've decided to leave it as a bit of a test of edge
behavior against it.