build: add ability to filter tests
Also remove anonymous struct for the acceptance tests, as it seems like
there's an accepted proposal to remove it from the language. We also
can't use it anymore for the acceptance tests as we need to add the
filter opts, and you can't do that after the fact on an anonymous
struct.