Surface: move to unmanaged and infallible design
This moves Surface and ImageSurface to an unmanaged, infallible design:

* Allocators are no longer stored in the surface.
* Anything that needs an allocator (e.g., init, deinit, and downsample)
  requires the same allocator to be used across the lifetime of the
  surface, as you would expect with unmanaged patterns within Zig.
* The only function that returns an error now is init. Every other
  function has a well-defined failure mode. Particularly, getPixel will
  return null on out-of-bounds requests, and setPixel will no-op.
  Additionally every pixel is valid on a surface for setPixel and
  paintPixel, using copySrc now instead of fromPixel (essentially making
  mismatched pixel types a cast).

Additionally, ImageSurface has gotten an initBuffer initializer. Using
this function, the only sanity checking that's done is that you're using
non-negative values for width and height; negative values cause a panic
instead of an error. Otherwise, you are expected to have supplied the
appropriately sized buffer for the appropriate type. There is a
counterpart downsampleBuffer to use when you use this
bring-your-own-buffer approach, so you don't have to supply an allocator
to anything else (as if you did, it would be an illegal operation).

Note that due to the current state of the language and ptrCast-ing
between slices of different alignments, initBuffer is currently not
exposed in the union interface. As such, if you want to use initBuffer,
you have to call the type function directly and then use
asSurfaceInterface.