Path: move to unmanaged design, introduce StaticPath, other improvements
This commit changes Path to be a fully unmanaged struct, which
essentially means changing its backing ArrayList to ArrayListUnmanaged.
All general calls to Path now require an allocator, which is then passed
to the ArrayListUnmanaged calls.

In addition to this work, however, we have broken off a number of the
calls to have AssumeCapacity counterparts, and also added initBuffer.
Most of these calls are completely infallible (not just memory
infallible, since they no longer require checking for allocation
errors); the only fallible calls that remain are the ones that require a
current point.

This also does a couple of other things:

* arc has now been split into arc and arcNegative, similar to Cairo -
  one draws clockwise, the other draws counter-clockwise.

* tolerance for arc calculations have now been removed from the arc
  calls., instead being a setting in the Path itself.

Both of these things vastly simplifies the use of the arc calls, and
sets the stage for further refactoring (specifically, the Context will
be managing synchronizing tolerance between it and its embedded Path).

Additionally, Path.close has had an ineffectual check for an initial
point moved to an assert. We had this put in an else clause when
capturing the optional before, but since the refactor and a move to
including the AssumeCapacity counterparts, this has proven unnecessary.
This makes closeAssumeCapacity completely infallible.

Finally, we have introduced a *completely infallible* interface to Path
called StaticPath. This type function wraps a buffer of a comptime-known
node length, with functions that call out to the AssumeCapacity
counterparts. Any of the calls that do require a current point are
safety-checked undefined behavior when they are called without a current
point.