internal/Painter: remove configurable edge FBA, use fallback alloc
This removes the configurable FBA for the edge cache, moving it to a
fallback allocator instead, with the FBA being a much smaller size, so
that we respect the stack a lot more.

The FBA should be enough for 64 edges worst-case scenario each edge is 4
bytes and may be duplicated in the final edge set, so 64*4*2 =
512 bytes. Note that not all edges make it into the final edge set
(e.g., during non-zero fill), hence why we say worst-case. This can and
will be optimized eventually to allow for more.