painter: new public and unmanaged interface
This makes the public painter interfaces public, and setting them up so
that they can be used in an unmanaged way.

The idioms of the new "fill" and "stroke" operations are set up so that
take take a required allocator, surface, pattern, and nodes to plot.
Additionally, an options struct is supplied to both which contain
optional arguments for both operations, with defaults that are in line
with what is currently available to both in the Context. This allows for
default behavior of these various operations should the user wish to do
so.