painter: use reference to Pattern in args
This is a pretty minimal change that brings Pattern idioms in line with
Surface idioms.

Since we have not implemented anything else other than single-pixel
surfaces yet, this is a pretty simple change; OpaquePattern is already
not a reference, so all that's needed adjusting is the args in the
painter functions, and setting getPixel to not return an error.

The getPixel API may eventually change depending on what the expectation
will be for patterns for returning pixels potentially out of bound from
a pattern that represents something with boundaries, such as a surface.