Context: new managed interface
This commit is the final piece of the current API overhaul - the managed
Context.

This is now the place where all managed drawing operations are expected
to take place. It pulls together all of our umnanaged components -
surfaces, patterns, paths, and painter operations, and provides a
managed frontend to these operations, doing things such as holding an
allocator for said unmanaged operations (that need one), and providing
synchronization of settings such as transformation matrices and
tolerance.

Helpers have been added to get and set every setting within the context;
it's now considered unsupported behavior for a consumer to set a field
directly on a context (or need to set one directly, for that matter,
which should help steer further updates).

If someone requires functionality that the Context is incapable of
providing or would be considered unsuitable for the context, the
recommendation is to now use the individual components in an unmanaged
way. For this, the Context serves as an example for how this can be
done.