Make all error sets explicit
This makes all returned error unions in the API explicit, setting the
precedent that we will expect this in the future.

I tried to follow the following pattern, with exceptions for semantics
where I thought it made sense:

* No more central error package - error sets are located within their
  own appropriate type or package, see below.

* Each package or type, where appropriate, got its own "Error" set that
  *only* has local errors in it (i.e., specific named errors returned,
  not passed up from a "try" predicate).

* Any error sets from the standard library (std.mem.Allocator.Error being
  by far the most common) are merged with these error sets at the site
  of the function signature. This allowed for differentiation -
  functions that only returned std.mem.Allocator.Error, for example,
  could just have that listed, and vice versa, or both, if need be. This
  makes it easy in a lot of cases (especially Path) to say "this call
  will not return any memory errors", etc.

Exceptions to the above (especially with regards to merging at the
function signature) do exist where it would make sense for readability,
good examples are the painter package and the PNG export functions; the
latter especially merges a *lot* of error sets as it does extensive file
I/O and compression, so there are a lot of sets to traverse.

Fixes #48.