export_png: allow exporting of alpha surfaces to greyscale
This allows the support of exporting of alpha8 pixel surfaces to
greyscale. This is in line with Cairo.

A pretty simple change actually, given that we have most of the other
things in place for PNG export, we just add it as a supported surface
type and handle it in header and pixel encoding.