Introduce 1, 2, and 4-bit alpha formats
This introduces pixel formats for 1, 2, and 4-bit alpha-channel.

Using these formats as a surface allows you to take advantage of an
aptly packed backing ImageSurface, generated by the new
PackedImageSurface type function. This packs the buffer using a
PackedIntSliceEndian in little-endian fashion, which is consistent with
our packed pixel formats. On buffer size, this ultimately means that you
could save up to 8 times the buffer space if you just need a 1-bit alpha
surface instead of alpha8 (or 24 or 32 times savings if you use the PNG
export for greyscale before ae6cb42.

During anti-aliasing, our mask surface also matches our target surface
if it is below 8-bit alpha. Note that we actually skip anti-aliasing for
1-bit alpha as it's essentially pointless, so if you use 1-bit alpha you
will get the same results as you would with AA off. Note to take
advantage of the compositing surface memory optimizations when rendering
to 2, 4, or 8-bit alpha surfaces, your pixel source needs to match the
target format; this may change in the future where we just pre-cast the
alpha source to the mask to avoid the RGBA surface allocation.

As part of this work as well we've moved all alpha pixel formats to be
generated by a generic type function to avoid repeating the same logic 3
or 4 times.