internal: polygon plotting memory tweaks
This makes the following tweaks to memory allocation during polygon
corner plotting and edge finding:

* Removes the ArenaAllocator from Polygon. Instead, all polygon corners
  are allocated/de-allocated as needed using an unmanaged pattern,
  one-by-one. This makes allocations easier to observe and is a trivial
  change with regards to implementation of de-allocation, which was the
  point originally.

* Switches PolygonList from using an ArrayList to a SinglyLinkedList.
  The main purpose here is two-fold. First, it removes the standard
  over-allocation that ArrayList does; this is (current_len / 2 + 8)
  every time an expansion needs to happen, which can lead to excessive
  amounts of memory allocated versus what is needed. Second, it removes
  the need for re-allocations when adding new items to the list,
  something that would still happen if we were doing something such as
  using ensureCapacityPrecise on top of append to avoid over-allocation.

* Sets initial capacities for our edge FBA that's used during the
  scanline rasterization. This is based on the findings from above - we
  want to make sure each tier on the FBA stack is allocated properly to
  avoid re-allocations that would ultimately create wasted space; only
  space at the top of the stack is re-used on an FBA when it's freed,
  so we need to make sure the total edge list - the list that's on the
  bottom of the stack - is appropriately sized.

Note that the first two changes for sure result in slower render times,
likely due to the much larger amount of allocations that are happening.
I'm considering this an acceptable change - for the time being, as I
want to focus on memory optimizations to ensure as little RAM overhead
in rendering as possible. Eventually we will address speed, possibly
with something such as specialized allocators for the whole of polygon
plotting that can address scenarios where memory is not so much of a
concern.