internal: move all ArrayLists to be unmanaged
There were a couple of ArrayLists that were set up to be managed that
didn't need to be, mainly in tests but also in Pen. This has been fixed
now.

Also removed Pen.verticesFor as we are not using it, requires
allocation, and we were storing the allocator for it - oops!