internal: don't initialize the pen on stroke if we don't need it
This changes the stroke plotter to only init the pen if we need it,
saving us allocations and time for operations that don't use round caps
or joins.

Since all beziers are round-joined but it's costly to determine if there
was a bezier in the path, we also lazy-init the pen on the first
curve_to encountered. We store the pen in an optional now so it's easy
to see if the vertices need to be freed on deinit.