stroke: add dashed lines support
This adds dashed line support to our stroke operation.

These operations are controlled by a couple of options:

* "setDashes" (or "dashes" in the unmanaged painter) sets a slice of
  dash lengths that alternate between drawing a dash and spacing a gap.
  Single or odd-length slices are supported, in this case the dasher
  will wrap around, giving same-length dashes and gaps for single
  element slices and more complex patterns otherwise. More complex
  scenarios are supported, see the help on the context for further
  details.

* "setDashOffset" (or "dash_offset" in the unmanaged painter) sets the
  dash offset, pulling (positive value, i.e., fast-forwarding) or
  pushing (negative value, i.e., re-winding) the dasher state depending
  on the offset. This can be used to better align a dash pattern on a
  shape.

An empty or invalid slice passed to setDashes will disable dashing.

As part of this work, the stroke plotter got a rewrite to make it easier
to understand. The node iterator is gone, as is the optional-driven
state. The latter has been replaced by a buffer that allows the first
two initial points in a path to be saved idempotently while walking
through a path of 2+ nodes where those points need to be retained, but
we still need the most recent three nodes for a join. Some of the
finalization functions for appending polygons have also been made
generic to work with the dashed stroke plotter. All in all, the hope is
that the plotter has been made a lot simpler to be able to understand
the fundamentals of the non-dashed version, from which going on to the
dashed version should be easy to do.