Path: ensure relative helpers work in user space
This fixes the relative helpers to ensure that they operate in user
space versus device space by transforming the current point in the path
helper back to user space before applying the relative co-ordinates.

This currently was not happening, so the transformation was being
re-applied to coordinates that were in device space already, layering
the transformation on multiple times, which is definitely not what we'd
expect or want.