Make Tracy Profiler a lazy dependency
This makes Tracy a lazy dependency so that neither zig-tracy nor the
Tracy source are pulled in when they're not needed (normal use of the
library, running unit tests, running E2E tests with Tracy disabled).

Some conditional importing and code adjustments needed to be made to the
E2E tests to make this happen to ensure that zig-tracy was only imported
when it was going to be available.