Dasher: fix/spin off validation logic
There was an issue in the validation logic that would have allowed the
full-zero case to go through. This has now been fixed (there has to be
at least one greater-than-zero element present in order to be valid,
with no negative values).

The validation logic has also been spun off from init so that the dasher
is initialized in the dash plotter versus the generic stroke plotter.
This was a minor issue that I missed fixing before 0.5.0 - it should
avoid a minor amount of stack overhead and a copy, not a huge deal, but
these are things I'm trying to be more cognizant of going forward.