gradient: add linear gradients
This adds linear gradients as the next pixel source (pattern) for z2d.

This initial effort offers the ability to do multi-stop gradients
specified in device space (so bounded to the destination surface either
in the context or when using lower-level composition). Our color
function (for the time being) is simple linear interpolation, although
other color functions will follow operating in other colorspaces and
providing the ability to do dithering to prevent banding.

It's (currently) expected that folks will work with the gradient package
directly to manage gradients and then set them in the context with
something like:

    context.setSource(gradient.asPatternInterface());

Although this may change. As such, we have not included much extension
to the Pattern interface at this time, only opting to add it to the
union and ensure that getPixel is supported.

As part of this work, "Context.setSource" as it exists currently has
been changed to "Context.setSourcePixel" to make room for setting
general source patterns. "Context.getSource" now returns the pattern
union, not the Pixel.