color: introduce high-level interface for color spaces
This introduces an MVP for a high-level interface for color spaces, or a
way for providing color that does not need to be provided as raw pixel
data.

The following color spaces are being provided initially: linear RGB,
sRGB, and HSL. More will follow after other v0.6.0 deliverables are
complete (possibly after v0.6.0).

A major part of this is the ability to perform interpolation for
gradients in non-RGB space - e.g., in HSL, using a particular method for
traversing the color circle (shortest path, longest path, etc). As such,
gradients have also been changed to work off of colors versus direct
pixel data. Some quality-of-life shorthand is also used to allow for the
ability to provide color space data as a "verb" of arguments, versus
long composition chains to wrap every type (this has also been added to
Pixel). We will eventually expand the shorthand to things like the named
CSS color list, etc.

This commit should also open the doors up for us to properly
gamma-correct PNG exports. This has not been done in this commit to keep
the changeset small, and as such will follow in future commits.

A small note about gamma: currently only fast-gamma is provided, but the
intention is to provide the ability to use the more sophisticated full
transfer functions in later iterations.